スポットリスト取得処理を微修正
